Move file struct to its own file; rename to File
